Chore: refactor STAC filter test
